requirements.txt
pyodbc
apscheduler
fastapi
uvicorn
python-dotenv
pytelegrambotapi
sqlalchemy
oauth2client
google-api-python-client
